Use simpler, but equivalent expression
These are equivalent. The original assignment uses 3 args, one which
expands to use the other:

106       this.x = x - this.xCenter - this.xDistanceToCenter

expanded with these that are set in fingerDown:

125       this.xStart = x
126       this.yStart = y
127       this.xCenter = event.srcElement.clientWidth / 2
128       this.xDistanceToCenter = x - this.xCenter

becomes:

this.x = x - this.xCenter - (this.xStart - this.xCenter)

then collapses to:

this.x = x - this.xStart

IMO this is simpler and makes it clearer that this.x is the x TRAVEL.

If we wanted to be super clear for future readers we might change the
variable name to this.xDelta, x.Travel, or something like that.