Remove xStart offset
Since this.x is shown from the last commit to already be the *travel*,
that means it already contains the comparison to this.xStart.

Therefore we don't want to compare it again.

The operation we want to do is see if the TRAVEL is greater than the
minSwipeDistanceBeforeAction,

not see if the travel is greater than the start PLUS min activation
distance.

-        if (this.x > this.xStart + this.minSwipeDistanceBeforeAction) {
+        if (this.x > this.minSwipeDistanceBeforeAction) {