Allow dragging through start point
Original logic ALWAYS skipped card rotation logic if the drag distance
was within the sensitivity range from it's start position.

This means that if the card is picked up in the center, then dragged
right, then brought BACK through it's original point, it would not
display smooth drag behavior through it's original start point.

This commit:

1. makes it so that subsequent motion through the origin drags smoothly
2. makes it so the menu hiding behavior happens on visible UI drag
start, rather than as soon as the finger is down